Handle zero-length arrays
In mangled names, ulam uses an array length of 0 ('10' in lex) as
a flag meaning 'this is a scalar rather than an array'. But since
zero length arrays of classes or primitives are legal, ulam uses
'n11' (leading 'n' as a(noften unimplemented) notation for negative
numbers) to specify an array length of -1. But the MFM core code
wasn't prepared to parse (or generate) mangled names using that
convention. (This would cause failures trying to display a
value containing a zero length array in an Atom View Panel, for
example.) Now the MFM core code handles this case.